feat(config)!: remove the K_ prefix, the --run-* flags and the bump alias
ADR 0011 made the config file the contract and kept every older spelling
working beside it, so an installation could migrate one Deployment at a
time. That migration is what the deprecations were for, and holding them
means configuration is read twice over: a component selected by the
`services` section *or* by a flag, a top-level key answered by the file
*or* by a K_ variable, with a ranking that had to be argued and tested.

All three are removed. A --run-* flag is unregistered rather than hidden,
so cobra refuses it and the process exits non-zero — a flag that still
parsed and selected nothing would give a pod that comes up healthy running
none of what its manifest asked for. The startup warning naming every K_
variable in the environment stays, saying "no longer read" instead of
"deprecated": the failure mode of the removal is a setting that is simply
absent, and this line is what names the cause. A variable the file itself
refers to is still left out of it.

x/config now reads no configuration value through a viper accessor, so
ADR 0011's boundary holds without its exception list, and the promotion of
a K_ variable onto its key — with the nested merge it needed in order not
to hide the keys beside the one it wrote — goes with it.

See ADR 0012.
