+++ x/config/services_test.go
@@ -14,8 +14,9 @@ func assertEnabled(t *testing.T, got Services, want ...string) {
 	}
 }
 
-// Nothing runs unless something says so, and an absent section is not an error:
-// a deployment migrating one component at a time still has flags for the rest.
+// Nothing runs unless the file says so, and an absent section is not an error
+// here: the boot path is what refuses a process that would run nothing, with a
+// message naming the section rather than a decode failure.
 func TestLoadServices_NothingByDefault(t *testing.T) {
 	viper.Reset()
 	t.Cleanup(viper.Reset)
@@ -56,66 +57,23 @@ services:
 	assertEnabled(t, services, "stats", "api")
 }
 
-// The deprecated flag and the section are OR-ed, so a pod still passing --run-api
-// keeps working next to one that has moved to the file.
-func TestLoadServices_DeprecatedFlagsStillEnable(t *testing.T) {
+// The removed keys are read by nothing: the `services` section is the whole
+// answer, and a `run-api` key surviving somewhere in viper does not add a
+// component behind the file's back.
+func TestLoadServices_IgnoresTheRemovedRunKeys(t *testing.T) {
 	viper.Reset()
 	t.Cleanup(viper.Reset)
 
 	viper.Set("services.stats.enabled", true)
-	viper.Set("run-api", true)
-
-	services, err := LoadServices()
-	if err != nil {
-		t.Fatalf("LoadServices: %v", err)
+	for _, key := range []string{"run-api", "run-verifier", "run-bounce"} {
+		viper.Set(key, true)
 	}
-	assertEnabled(t, services, "stats", "api")
-}
-
-// A flag saying nothing does not turn off what the file turned on: false is the
-// flag's default, not an instruction.
-func TestLoadServices_AFalseFlagDoesNotDisable(t *testing.T) {
-	viper.Reset()
-	t.Cleanup(viper.Reset)
-
-	viper.Set("services.stats.enabled", true)
-	viper.Set("run-stats", false)
 
 	services, err := LoadServices()
 	if err != nil {
 		t.Fatalf("LoadServices: %v", err)
 	}
 	assertEnabled(t, services, "stats")
-}
-
-// Regression, from when these were flags: --run-bounce / --run-verifier must be
-// promoted onto their canonical names before the services section is read, or the
-// alias fires too late, nothing is registered, and the process exits as if it had
-// been asked to run nothing.
-func TestLoadServices_DeprecatedAliasesArePromotedFirst(t *testing.T) {
-	for _, tc := range []struct {
-		alias string
-		want  string
-	}{
-		{alias: "run-bounce", want: "tracker"},
-		{alias: "run-verifier", want: "validator"},
-	} {
-		t.Run(tc.alias, func(t *testing.T) {
-			viper.Reset()
-			t.Cleanup(viper.Reset)
-
-			viper.Set(tc.alias, true)
-
-			if _, err := Read(); err != nil {
-				t.Fatalf("Read: %v", err)
-			}
-			services, err := LoadServices()
-			if err != nil {
-				t.Fatalf("LoadServices: %v", err)
-			}
-			assertEnabled(t, services, tc.want)
-		})
-	}
 }
 
 // A misspelled service name is refused rather than ignored. Silently ignoring it
@@ -154,15 +112,4 @@ func TestLoadServices_ReportsAnUnresolvableReference(t *testing.T) {
 func TestAllServices(t *testing.T) {
 	assertEnabled(t, AllServices(),
 		"sender", "dispatcher", "validator", "stats", "tracker", "api", "smtp", "audit")
-}
-
-// The name in the file, the name of the deprecated flag and the name in a log
-// line are one string, and this is what says so.
-func TestDeprecatedRunKeyMatchesEveryService(t *testing.T) {
-	var s Services
-	for _, svc := range s.each() {
-		if got, want := deprecatedRunKey(svc.name), "run-"+svc.name; got != want {
-			t.Errorf("deprecatedRunKey(%q) = %q, want %q", svc.name, got, want)
-		}
-	}
 }
